requirements.txt
python-dotenv
google-genai
google
requests